git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@184997 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/g++.dg/tree-ssa/tmmti-2.C
@@ -3,7 +3,7 @@
 
 int a[4][8];
 
-int foo(int i)
+int foo(long i)
 {
 	return *(&a[0][0] + i*8); // a[i][0]
 }
@@ -12,11 +12,11 @@ struct Foo { double x, y; };
 
 Foo b[4];
 
-double bar(int i)
+double bar(long i)
 {
 	return *(&b[0].x + i*2); // b[i].x
 }
 
-/* { dg-final { scan-tree-dump "a\\\[.*i.*\\\]\\\[0\\\]" "optimized" } } */
+/* { dg-final { scan-tree-dump "a\\\[.*i.*\\\]\\\[0\\\]" "optimized" { xfail *-*-* } } } */
 /* { dg-final { scan-tree-dump "b\\\[.*i.*\\\].x" "optimized" } } */
 /* { dg-final { cleanup-tree-dump "optimized" } } */
