+++ gcc/testsuite/gfortran.dg/g77_intrinsics_sub.f
@@ -1,7 +1,9 @@
 ! { dg-do compile }
+! { dg-options "-std=legacy" }
+!
 ! Testing g77 intrinsics as subroutines
-      integer*8 i8, j8
-      integer*4 i4, j4
+      integer(kind=8) i8, j8
+      integer i4, j4
       integer i, j
       character*80 c
